Extract top-level functions to helpers, add tests
