+++ controllers/apply-next/form-router.js
@@ -1,6 +1,6 @@
 'use strict';
-const { cloneDeep, concat, find, flatMap, isEmpty, pick, set, unset } = require('lodash');
-const { get, getOr } = require('lodash/fp');
+const { concat, flatMap, isEmpty, set, unset } = require('lodash');
+const { get } = require('lodash/fp');
 const { matchedData } = require('express-validator/filter');
 const { validationResult } = require('express-validator/check');
 const express = require('express');
@@ -11,164 +11,7 @@ const cached = require('../../middleware/cached');
 const { requireUserAuth } = require('../../middleware/authed');
 const ApplicationService = require('../../services/applications');
 const { localify } = require('../../modules/urls');
-
-function prepareForm(locale, formModel, formData) {
-    const localise = get(locale);
-
-    const translateField = field => {
-        field.label = localise(field.label);
-        field.explanation = localise(field.explanation);
-
-        const match = find(formData, (value, name) => name === field.name);
-        if (match) {
-            field.value = match;
-        }
-
-        // Translate each option (if set)
-        if (field.options) {
-            field.options = field.options.map(option => {
-                option.label = localise(option.label);
-                option.explanation = localise(option.explanation);
-                return option;
-            });
-        }
-        return field;
-    };
-
-    const translateStep = step => {
-        step.title = localise(step.title);
-
-        step.fieldsets = step.fieldsets.map(fieldset => {
-            fieldset.legend = localise(fieldset.legend);
-            fieldset.introduction = localise(fieldset.introduction);
-            fieldset.fields = fieldset.fields.map(translateField);
-            return fieldset;
-        });
-
-        return step;
-    };
-
-    const translateSection = section => {
-        section.title = localise(section.title);
-        if (section.summary) {
-            section.summary = localise(section.summary);
-        }
-        if (section.introduction) {
-            section.introduction = localise(section.introduction);
-        }
-        return section;
-    };
-
-    const clonedForm = cloneDeep(formModel);
-    clonedForm.title = localise(formModel.title);
-
-    clonedForm.sections = clonedForm.sections.map(section => {
-        section = translateSection(section);
-        section.steps = section.steps.map(translateStep);
-        return section;
-    });
-
-    if (clonedForm.termsFields) {
-        clonedForm.termsFields = clonedForm.termsFields.map(translateField);
-    }
-
-    if (clonedForm.titleField) {
-        clonedForm.titleField = translateField(clonedForm.titleField);
-    }
-
-    return clonedForm;
-}
-
-const FORM_STATES = {
-    incomplete: {
-        type: 'incomplete',
-        label: 'In progress'
-    },
-    complete: {
-        type: 'complete',
-        label: 'Complete'
-    },
-    invalid: {
-        type: 'invalid',
-        label: 'Invalid'
-    },
-    empty: {
-        type: 'empty',
-        label: 'Not started'
-    }
-};
-
-function getFieldsForStep(step) {
-    return flatMap(step.fieldsets, fieldset => fieldset.fields);
-}
-
-function getFieldsForSection(section) {
-    return flatMap(section.steps, getFieldsForStep);
-}
-
-function getAllFields(formModel) {
-    return flatMap(formModel.sections, getFieldsForSection);
-}
-
-/**
- * Build up a set of state parameters for steps, sections, and the form itself
- * eg. to show status on the summary page
- */
-function validateFormState(form, formData, sessionValidation) {
-    const clonedForm = cloneDeep(form);
-
-    clonedForm.sections = clonedForm.sections.map(section => {
-        section.steps = section.steps.map((step, stepIndex) => {
-            // Handle steps that this user doesn't need to complete
-            if (step.matchesCondition && step.matchesCondition(formData) === false) {
-                step.state = FORM_STATES.complete;
-                step.notRequired = true;
-            } else {
-                const fieldsForStep = getFieldsForStep(step);
-                const stepData = pick(formData, fieldsForStep.map(field => field.name));
-
-                if (isEmpty(stepData)) {
-                    step.state = FORM_STATES.empty;
-                } else {
-                    // @TODO construct this via a function
-                    const stepIsValid = getOr(false, `${section.slug}.step-${stepIndex}`)(sessionValidation);
-                    step.state = stepIsValid ? FORM_STATES.complete : FORM_STATES.incomplete;
-                }
-            }
-
-            return step;
-        });
-
-        // See if this section's steps are all empty
-        const sectionIsNotEmpty = section.steps
-            .filter(step => !step.notRequired)
-            .some(step => step.state !== FORM_STATES.empty);
-
-        if (sectionIsNotEmpty) {
-            // Work out this section's state based on its steps' state
-            const sectionHasInvalidSteps = section.steps
-                .filter(step => step.notRequired === undefined)
-                .some(step => step.state !== FORM_STATES.complete);
-            section.state = sectionHasInvalidSteps ? FORM_STATES.incomplete : FORM_STATES.complete;
-        } else {
-            section.state = FORM_STATES.empty;
-        }
-
-        return section;
-    });
-
-    // Check whether the entire form is empty
-    const formIsNotEmpty = clonedForm.sections.some(section => section.state !== FORM_STATES.empty);
-
-    if (formIsNotEmpty) {
-        // Work out the entire form's state based on its sections' state
-        const formHasInvalidSections = clonedForm.sections.some(section => section.state !== FORM_STATES.complete);
-        clonedForm.state = formHasInvalidSections ? FORM_STATES.incomplete : FORM_STATES.complete;
-    } else {
-        clonedForm.state = FORM_STATES.empty;
-    }
-    return clonedForm;
-}
+const { FORM_STATES, findNextMatchingStepIndex, getAllFields, prepareForm, validateFormState } = require('./helpers');
 
 function initFormRouter(formModel) {
     const router = express.Router();
@@ -310,20 +153,14 @@ function initFormRouter(formModel) {
         sectionModel.steps.forEach((stepModel, stepIndex) => {
             const currentStepNumber = stepIndex + 1;
             const numSteps = sectionModel.steps.length;
-            const nextStep = sectionModel.steps[stepIndex + 1];
             const nextSection = formModel.sections[sectionIndex + 1];
 
             const fieldsForStep = flatMap(stepModel.fieldsets, 'fields');
             const validators = fieldsForStep.map(field => field.validator(field));
 
-            function redirectNext(req, res) {
-                /**
-                 * 1. If there a next step in the current section go there.
-                 * 2. Otherwise, if there's a next section, go there
-                 * 3. Otherwise, go to summary screen
-                 */
-                if (nextStep) {
-                    res.redirect(`${req.baseUrl}/${sectionModel.slug}/${currentStepNumber + 1}`);
+            function redirectNext(nextMatchingStepIndex, req, res) {
+                if (nextMatchingStepIndex !== -1 && nextMatchingStepIndex <= sectionModel.steps.length) {
+                    res.redirect(`${req.baseUrl}/${sectionModel.slug}/${nextMatchingStepIndex + 1}`);
                 } else if (nextSection) {
                     res.redirect(`${req.baseUrl}/${nextSection.slug}`);
                 } else {
@@ -332,7 +169,9 @@ function initFormRouter(formModel) {
             }
 
             function renderStep(req, res, errors = []) {
-                const sectionLocalised = res.locals.form.sections.find(s => s.slug === sectionModel.slug);
+                const { form, currentApplicationData } = res.locals;
+
+                const sectionLocalised = form.sections.find(s => s.slug === sectionModel.slug);
                 const stepLocalised = sectionLocalised.steps[stepIndex];
 
                 res.locals.breadcrumbs = concat(
@@ -346,18 +185,21 @@ function initFormRouter(formModel) {
                     }
                 );
 
-                if (
-                    stepModel.matchesCondition &&
-                    stepModel.matchesCondition(res.locals.currentApplicationData) === false
-                ) {
-                    redirectNext(req, res);
-                } else {
+                const nextMatchingStepIndex = findNextMatchingStepIndex(
+                    sectionModel.steps,
+                    stepIndex,
+                    currentApplicationData
+                );
+
+                if (nextMatchingStepIndex === stepIndex) {
                     res.render(path.resolve(__dirname, './views/step'), {
                         title: `${stepLocalised.title} | ${res.locals.form.title}`,
                         csrfToken: req.csrfToken(),
                         step: stepLocalised,
                         errors: errors
                     });
+                } else {
+                    redirectNext(nextMatchingStepIndex, req, res);
                 }
             }
 
@@ -376,7 +218,12 @@ function initFormRouter(formModel) {
                         // If a step has no errors, then mark it as valid
                         res.locals.setSessionData(validationPath, FORM_STATES.complete);
                         req.session.save(() => {
-                            redirectNext(req, res);
+                            const nextMatchingStepIndex = findNextMatchingStepIndex(
+                                sectionModel.steps,
+                                stepIndex + 1,
+                                res.locals.currentApplicationData
+                            );
+                            redirectNext(nextMatchingStepIndex, req, res);
                         });
                     } else {
                         // Remove this step from the valid list (if it was there before)
